pyproject.toml
[project]
name = "swara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lingua-language-detector>=2.1.1",
    "pandas>=2.3.2",
    "phonemizer>=3.3.0",
    "tqdm>=4.67.1",
]
